jobpipe/tailor/requirements.txt
anthropic
supabase
python-dotenv
playwright
apscheduler
httpx
beautifulsoup4
reportlab
pyyaml>=6.0
